Switch to Gtk.Application and Gtk.ApplicationWindow.
The Gtk.Application is the 'main' class of the app. The methods
do_startup() and do_activate() are called when you run the app. That's
 why I didn't connect anymore the "startup" add "activate" signals.
The Gtk.ApplicationWindow represents the main window of the application.
I had to create a separate class, because only loading the applicationwindow
from the Gtk.Builder doesn't display anything.