Switch to a non popup window for presenting a key.
Choosing a key now will display the details of it in a new page
of a Gtk.Notebook container, instead of opening a new window.
This is the old style we used before in gnome-keysign.

The 'KeyPresentPage' represents the class that loads
the second page in the notebook container. It doesn't
show its elements right now though, improvement must
be done.