Add a transition page after a fpr is entered.
It may happen that a key download gets stucked or takes too long
to complete. This transition page will try to offer some valuable
information in the future. For now, the transition page doesn't
do anything than displaying a spinner widget.

The idea that will be implemented in next commits, is that when
a mouse left click happens on the transition page, it will stop
the download, and right-click will advance to the "confirm to sign
this key" page.