Split the application into modules.
The major split is the Application class, into Application and
ApplicationWindow. This comes with a problem that I couldn't find
a full fix: the app had 2 windows. I temporary fixed the problem by
overwriting the ApplicationWindow.do_show_all() method, which is
called in the Application.do_activate() method.

This fixes the 2 windows problem, but I can't get the app
menu to be shown anymore.