Add scripts for building a flatpack app.
Use flatpack-build-app.sh to install gnome keysign as org.gnome.Keysign.
After that either run 'flatpack run org.gnome.Keysign' or use the
supplied Makefile.

The app will not work and that's because we're using a python file
and I guess something else must be done. I've tried with wrapping
the gnome-keysign.py file in a start-keysign.sh script, but no
success.