Add setup.py file and move app files to keysign package.
This makes possible to install the app via pip install.
It tries to load the UI files from the installation dir, and if
fails, than it loads them from current 'data' dir. The reason
for that is because in this development stage, we need to test
the app many times, and to install it each time it would be
too much of a hassle.