Use the new KeyFprScanWidget widget.
This is really wierd, the camera won't show anything even now. It
might not like the containerization, but I don't realy now.
The KeyFprScanWidget works, if called alone, but somehow, while
called within the app, the camera display is blank, although it can
scan a barcode if put in front of it.